Change keyValues to items per change in `pythonic`
