+++ cli.js
@@ -15,8 +15,21 @@ settings.dbname = program.dbname || settings.dbname;
 settings.dbhost = program.dbhost || settings.dbhost;
 settings.timeout = program.timeout || settings.timeout;
 
-const {app} = require('./dist');
+const {app, agenda} = require('./dist');
 
-app.listen(program.port, () => {
+const server = app.listen(program.port, () => {
   console.log(`App listening on port ${program.port}.`);
 });
+
+const graceful = () => {
+  console.log('\nShutting down gracefully...');
+  agenda.stop(() => {
+    console.log('\nClosing server...');
+    server.close();
+    // eslint-disable-next-line unicorn/no-process-exit
+    process.exit(0);
+  });
+};
+
+process.on('SIGTERM', graceful);
+process.on('SIGINT', graceful);
